выполнение ДЗ

+++ src/main/java/com/prosky/homeworktesting2/service/CalculatorService.java
@@ -12,13 +12,11 @@ public class CalculatorService {
         return num1 + num2;
     }
 
-    public int minusCalculator(int num1, int num2) {
+    public double minusCalculator(double num1, double num2) {
         return num1 - num2;
     }
 
-    //    public String multiplyCalculator(int num1, int num2) {
-    public int multiplyCalculator(int num1, int num2) {
-//        return String.valueOf(num1 * num2);
+    public double multiplyCalculator(double num1, double num2) {
         return num1 * num2;
     }
